refactor: move rate limit and session constants to a new server-constants file
- Transferred RATE_LIMIT and SESSION constants from constants.ts to a newly created server-constants.ts file for better organization and separation of concerns.
- Updated import paths in auth.ts to reflect the new location of these constants.
